Passing filename, not path, to pyreadr
pyreadr only accepts strings, not path objects, for it's
file names - so have had to extract the filename string to
make this interface work.

(I think this way is better than passing the string, because
using Path gives us the checks to ensure the file exists,
without having to mess about with other libraries).