Changes to AURN reference and required station listing
Moving towards standardised methods for AURN and
met processing.

Added function for organising the calculation of the
lists of required and reference sites - but made it more
generic and flexible than the met function.

Adapted the determination of which stations to process,
this will now be the intersection of the requested station list
and the list of required stations.

Adapted the calculation of how many reference sites to use,
this is based on the min of 5 / how many reference sites there
are for each of the species of interest.